requirements.txt
streamlit
torch
langchain
langchain-community
pypdf
python-dotenv
sentence-transformers
transformers
accelerate